requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
markdown
pyyaml==6.0.2
